merge

+++ backend/User.php
@@ -10,11 +10,28 @@ class User
         $this->conn = $conn;
     }
 
-    private function generateAccessToken($userId) {
-        return md5($userId . time()); 
+    private function generateAccessToken($userId)
+    {
+        return md5($userId . time());
     }
-    
-    public function getAllUsers()
+
+    public function checkAcessToken($accessToken)
+    {
+        $query = "SELECT id, administrator FROM users WHERE access_token = ?";
+        $stmt = $this->conn->prepare($query);
+        $stmt->bind_param("s", $accessToken);
+        $stmt->execute();
+        $stmt->bind_result($id, $administrator);
+        $stmt->fetch();
+        $stmt->close();
+        if ($id !== null && $administrator !== null) {
+            return array('id' => $id, 'role' => $administrator ? 'admin' : 'user');
+        } else {
+            return false;
+        }
+    }
+
+    public function getAllUsers($data)
     {
         /* GET ALL USERS + QUESTIONS + ANSWERS
         $query = "SELECT users.username, users.password, users.administrator, questions.date, 
@@ -25,6 +42,15 @@ class User
         LEFT JOIN answers ON answers.question_id = questions.id
         LEFT JOIN types ON answers.type_id = types.id";
         */
+        $accessToken = $data['access_token'];
+        $accessTokenData = $this->checkAcessToken($accessToken);
+        if (!$accessTokenData) {
+            return false;
+        }
+        if ($accessTokenData['role'] !== "admin") {
+            return false;
+        }
+
         $query = "SELECT id, username, administrator FROM users";
         $result = mysqli_query($this->conn, $query);
         $users = [];
@@ -38,8 +64,7 @@ class User
     {
         $username = $data['username'];
         $password = $data['password'];
-        $query = "SELECT id, username, password, administrator FROM users
-              WHERE username = ?";  
+        $query = "SELECT id, username, password, administrator FROM users WHERE username = ?";
         $stmt = $this->conn->prepare($query);
         $stmt->bind_param("s", $username);
         $stmt->execute();
@@ -48,18 +73,38 @@ class User
         $stmt->close();
         if ($user && password_verify($password, $user['password'])) {
             $accessToken = $this->generateAccessToken($user['id']);
-            $user['access_token'] = $accessToken;    
-            unset($user['password']); 
+            $stmt = $this->conn->prepare("UPDATE users SET access_token = ? WHERE username = ?");
+            $stmt->bind_param("ss", $accessToken, $username);
+            $stmt->execute();
+            $stmt->close();
+            $user['access_token'] = $accessToken;
+            unset($user['password']);
             return $user;
         } else {
-            return null; 
+            return null;
         }
+    }
+    public function logout($data)
+    {
+        $accessToken = $data['access_token'];
+        $query = "UPDATE users SET access_token = NULL WHERE access_token = ?";
+        $stmt = $this->conn->prepare($query);
+        $stmt->bind_param("s", $accessToken);
+        $stmt->execute();
+        if ($stmt->affected_rows > 0) {
+            $stmt->close();
+            return true;
+        } else {
+            $stmt->close();
+            return false;
+        }
+
     }
 
     public function checkUsername($data)
     {
         $username = $data['username'];
-        $query = "SELECT id FROM users WHERE username = ?";  
+        $query = "SELECT id FROM users WHERE username = ?";
         $stmt = $this->conn->prepare($query);
         $stmt->bind_param("s", $username);
         $stmt->execute();
@@ -67,7 +112,7 @@ class User
         $user = $result->fetch_assoc();
         $stmt->close();
         if ($user) {
-            return true; 
+            return true;
         } else {
             return false;
         }
@@ -78,16 +123,15 @@ class User
         $username = $data['username'];
         $password = $data['password'];
         $hashed_password = password_hash($password, PASSWORD_ARGON2ID);
-        $administrator = $data['administrator'];        
-        $stmt = $this->conn->prepare("SELECT id FROM users WHERE username = ? AND password = ?");
-        $stmt->bind_param("ss", $username, $hashed_password);
+        $stmt = $this->conn->prepare("SELECT id FROM users WHERE username = ?");
+        $stmt->bind_param("s", $username);
         $stmt->execute();
         $result = $stmt->get_result();
         if ($result->num_rows > 0) {
             return false;
         }
-        $stmt = $this->conn->prepare("INSERT INTO users (username, password, administrator) VALUES (?, ?, ?)");
-        $stmt->bind_param("ssi", $username, $hashed_password, $administrator);
+        $stmt = $this->conn->prepare("INSERT INTO users (username, password, administrator) VALUES (?, ?, 0)");
+        $stmt->bind_param("ss", $username, $hashed_password);
         $stmt->execute();
         $result = $stmt->get_result();
         $stmt->close();
@@ -96,29 +140,49 @@ class User
 
     public function updateUser($id, $data)
     {
+        $accessToken = $data['access_token'];
         $username = $data['username'];
         $password = $data['password'];
-        $hashed_password = password_hash($password, PASSWORD_ARGON2ID);
-        $administrator = $data['administrator'];
-        $stmt = $this->conn->prepare("UPDATE users SET (username = ?, password = ?, administrator = ?) WHERE id = ?");
-        $stmt->bind_param("ssii", $username, $hashed_password, $administrator, $id);
-        $stmt->execute();
-        if ($stmt->affected_rows > 0) {
-            $stmt->close();
-            return true;
+
+        $accessTokenData = $this->checkAcessToken($accessToken);
+        if (!$accessTokenData || $this->checkUsername($data)) {
+            return false;
+        }
+        $role = $accessTokenData['role'];
+        if (($role === "admin") || (($role === "user") && ($id == $accessTokenData['id']))) {
+            $hashed_password = password_hash($password, PASSWORD_ARGON2ID);
+            $stmt = $this->conn->prepare("UPDATE users SET username = ?, password = ? WHERE id = ?");
+            $stmt->bind_param("ssi", $username, $hashed_password, $id);
+            $stmt->execute();
+            if ($stmt->affected_rows > 0) {
+                $stmt->close();
+                return true;
+            } else {
+                $stmt->close();
+                return false;
+            }
         } else {
-            $stmt->close();
             return false;
         }
     }
 
-    public function deleteUser($id)
-    {    
-        $stmt = $this->conn->prepare("DELETE FROM users WHERE id = ?");
-        $stmt->bind_param("i", $id);
-        $stmt->execute();
-        $affectedRows = $stmt->affected_rows;
-        $stmt->close();
-        return $affectedRows > 0;
-    }   
+    public function deleteUser($id, $data)
+    {
+        $accessToken = $data['access_token'];
+        $accessTokenData = $this->checkAcessToken($accessToken);
+        if (!$accessTokenData) {
+            return false;
+        }
+        $role = $accessTokenData['role'];
+        if (($role === "admin") || ($role === "user" && $id == $accessTokenData['id'])) {
+            $stmt = $this->conn->prepare("DELETE FROM users WHERE id = ?");
+            $stmt->bind_param("i", $id);
+            $stmt->execute();
+            $affectedRows = $stmt->affected_rows;
+            $stmt->close();
+            return $affectedRows > 0;
+        } else {
+            return false;
+        }
+    }
 }
